pyproject.toml
[tool.poetry]
name = "adventofcode2022"
version = "0.1.0"
description = ""
authors = ["Jesse van der Wolf <jessevanderwolf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
types-requests = "^2.28.11"
black = "^22.10.0"
autopep8 = "^2.0.0"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
